requirements_prod.txt
# NexVision IPTV — Production Dependencies
# Install: pip install -r requirements_prod.txt
# Updated: 2026-05-01 - v8.16 - Blueprint/service architecture + nav menu admin fix
# Entry points: app/wsgi.py (gunicorn) or run.py (dev)

# ── Core Flask ────────────────────────────────────────────────────────────────
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
blinker==1.9.0
click==8.3.1
flask-cors==6.0.2
PyJWT==2.12.1

# ── Production WSGI server ────────────────────────────────────────────────────
gunicorn==25.1.0
gevent==25.9.1              # async worker for 500+ concurrent connections
greenlet==3.3.2             # gevent dependency
zope.event==6.1
zope.interface==8.2

# ── MySQL (replaces sqlite3 in production) ────────────────────────────────────
PyMySQL==1.1.2
cryptography==46.0.7        # PyMySQL auth plugin support
cffi==2.0.0
pycparser==3.0

# ── Redis caching ─────────────────────────────────────────────────────────────
Flask-Caching==2.3.1
cachelib==0.13.0
redis==7.3.0

# ── Environment config ────────────────────────────────────────────────────────
python-dotenv==1.2.2

# ── HTTP client ───────────────────────────────────────────────────────────────
requests==2.33.0
urllib3==2.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11

# ── packaging ─────────────────────────────────────────────────────────────────
packaging==26.0

# ── Optional cloud storage (install separately if needed) ─────────────────────
# boto3                     # AWS S3 support
# azure-storage-blob        # Azure Blob support
# google-cloud-storage      # GCS support
